Combine all decorators into one fille
